commented out position change method
Changed the method and the objects, because the position was no longer an issue.
Method change is simple commented out a for loop that changed the position of the locks/blockers, now they stay were they were assigned in the scene instead of changing.
Objects panel scale was changed so that it is now correct instead of being adjusted for the position change.